Matts post picker modifications
